requirements.txt
flask==3.0.0
flask-cors==4.0.0
numpy==1.26.4
python-dotenv==1.0.0
openai==1.6.1
httpx==0.24.1
Pillow==10.2.0
requests==2.31.0
gunicorn==21.2.0